[pom] Add relativePath to force .m2 lookup
The root parent here will never be up one directory.  Its going to be in m2.  For users that have a root pom for other needs such as myself, this throws warnings as what is really being done is incorrect without this.

It should also be noted: oss-parent is obsolete and suggested against usage from sonatype.  It's best to pull in what is needed as there won't be any further updates.